+++ src/main/java/me/efekos/simpler/items/tag/AxolotlBucketTag.java
@@ -22,6 +22,7 @@
 
 package me.efekos.simpler.items.tag;
 
+@Deprecated
 public class AxolotlBucketTag extends ItemTag{
     private final Integer Variant;