+++ src/main/java/me/efekos/simpler/items/tag/BookTag.java
@@ -22,6 +22,7 @@
 
 package me.efekos.simpler.items.tag;
 
+@Deprecated
 public class BookTag extends ItemTag{
     public BookTag(ItemTag oldItemTag, String title, String author, Integer generation) {
         super(oldItemTag.Damage, oldItemTag.getHideFlags(), oldItemTag.display, oldItemTag.isUnbreakable(), oldItemTag.getCustomModelData(), oldItemTag.getEnchantments(), oldItemTag.RepairCost);